requirements.txt
openai~=1.60.2
ansible-core~=2.17.0
azure-search-documents
azure-core